# frontend/openapi.yaml
---
openapi: 3.0.3
info:
  title: backend API
  version: 1.0.0-SNAPSHOT
paths:
  /api/snapshot:
    post:
      tags:
      - Snapshot Controller
      operationId: createNewSnapshot
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                url:
                  type: string
                basic-info:
                  format: binary
                  type: string
                ss:
                  format: binary
                  type: string
                har:
                  format: binary
                  type: string
                console:
                  format: binary
                  type: string
                html:
                  format: binary
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewHarResponse'
  /api/snapshot/hello:
    get:
      tags:
      - Snapshot Controller
      operationId: hello
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
  /api/snapshot/{ref}:
    get:
      tags:
      - Snapshot Controller
      operationId: getShanpshotInfo
      parameters:
      - name: ref
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snapshot'
  /api/snapshot/{ref}/annotations-config:
    get:
      tags:
      - Snapshot Controller
      operationId: getSnapshotAnnotationsConfig
      parameters:
      - name: ref
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                format: binary
                type: string
  /api/snapshot/{ref}/annotations-svg:
    get:
      tags:
      - Snapshot Controller
      operationId: getSnapshotAnnotationsSvg
      parameters:
      - name: ref
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            image/svg+xml:
              schema:
                format: binary
                type: string
  /api/snapshot/{ref}/basic-info:
    get:
      tags:
      - Snapshot Controller
      operationId: getSnapshotBasicInfo
      parameters:
      - name: ref
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                type: object
  /api/snapshot/{ref}/console:
    get:
      tags:
      - Snapshot Controller
      operationId: getSnapshotConsoleLog
      parameters:
      - name: ref
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                type: object
  /api/snapshot/{ref}/finalize-capture:
    post:
      tags:
      - Snapshot Controller
      operationId: finalizeSnapshotCapture
      parameters:
      - name: ref
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                annotations-config:
                  format: binary
                  type: string
                annotations-svg:
                  format: binary
                  type: string
      responses:
        "201":
          description: Created
  /api/snapshot/{ref}/html:
    get:
      tags:
      - Snapshot Controller
      operationId: getSnapshotHtml
      parameters:
      - name: ref
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            text/html:
              schema:
                type: string
  /api/snapshot/{ref}/network:
    get:
      tags:
      - Snapshot Controller
      operationId: getSnapshotNetwork
      parameters:
      - name: ref
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                type: object
  /api/snapshot/{ref}/screenshot:
    get:
      tags:
      - Snapshot Controller
      operationId: getSnapshotScreenshot
      parameters:
      - name: ref
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            image/png:
              schema:
                format: binary
                type: string
components:
  schemas:
    Date:
      format: date
      type: string
      example: 2022-03-10
    NewHarResponse:
      type: object
      properties:
        id:
          type: string
    Snapshot:
      type: object
      properties:
        id:
          format: int64
          type: integer
        ref:
          type: string
        pageTitle:
          type: string
        pageUrl:
          type: string
        title:
          type: string
        description:
          type: string
        hasAnnotations:
          type: boolean
        created:
          $ref: '#/components/schemas/Date'
  securitySchemes:
    SecurityScheme:
      type: http
      description: Authentication
      scheme: basic